Refactor pandas fallback
